modules/admin/data/requirements.txt
PyYAML
python-ldap
